karaoke/workers/gpu/requirements.txt
minio
demucs
stable-ts
pydantic-settings
opencc
torchcodec